+++ core/src/main/java/nl/nn/adapterframework/align/ToXml.java
@@ -1,5 +1,5 @@
 /*
-   Copyright 2017 Nationale-Nederlanden
+   Copyright 2017,2018 Nationale-Nederlanden
 
    Licensed under the Apache License, Version 2.0 (the "License");
    you may not use this file except in compliance with the License.
@@ -76,6 +76,7 @@ public abstract class ToXml<C,N> extends XmlAligner {
 
 //	private boolean autoInsertMandatory=false;   // TODO: behaviour needs to be tested.
 	private boolean deepSearch=false;
+	private boolean failOnWildcards=false;
 
 	private String prefixPrefix="ns";
 	private int prefixPrefixCounter=1;
@@ -145,9 +146,13 @@ public abstract class ToXml<C,N> extends XmlAligner {
 	 */
 	public void startParse(C root) throws SAXException {
 		//if (DEBUG) log.debug("startParse() rootNode ["+node.toString()+"]"); // result of node.toString() is confusing. Do not log this.
-		validatorHandler.startDocument();
-		handleNode(root, getRootElement(), getTargetNamespace());
-		validatorHandler.endDocument();
+		try {
+			validatorHandler.startDocument();
+			handleNode(root, getRootElement(), getTargetNamespace());
+			validatorHandler.endDocument();
+		} catch (SAXException e) {
+			handleError(e);
+		}
 	}
 
 	public abstract N getRootNode(C container);
@@ -543,7 +548,28 @@ public abstract class ToXml<C,N> extends XmlAligner {
 			return true;
 		}
 		if (term instanceof XSWildcard) {
-			log.warn("getBestMatchingElementPath SHOULD implement term instanceof XSWildcard ["+ToStringBuilder.reflectionToString(term)+"]");
+			XSWildcard wildcard=(XSWildcard)term;
+			String processContents;
+			switch (wildcard.getProcessContents()) {
+			case XSWildcard.PC_LAX: processContents="LAX"; break;
+			case XSWildcard.PC_SKIP: processContents="SKIP"; break;
+			case XSWildcard.PC_STRICT: processContents="STRICT"; break;
+			default: 
+					throw new IllegalStateException("getBestMatchingElementPath wildcard.processContents is not PC_LAX, PC_SKIP or PC_STRICT, but ["+wildcard.getProcessContents()+"]");
+			}
+			String namespaceConstraint;
+			switch (wildcard.getConstraintType()) {
+			case XSWildcard.NSCONSTRAINT_ANY : namespaceConstraint="ANY"; break;
+			case XSWildcard.NSCONSTRAINT_LIST : namespaceConstraint="SKIP "+wildcard.getNsConstraintList(); break;
+			case XSWildcard.NSCONSTRAINT_NOT : namespaceConstraint="NOT "+wildcard.getNsConstraintList(); break;
+			default: 
+					throw new IllegalStateException("getBestMatchingElementPath wildcard.namespaceConstraint is not ANY, LIST or NOT, but ["+wildcard.getConstraintType()+"]");
+			}
+			String msg="term for element ["+baseElementDeclaration.getName()+"] is WILDCARD; namespaceConstraint ["+namespaceConstraint+"] processContents ["+processContents+"]. Please check if the element typed properly in the schema";
+			if (isFailOnWildcards()) {
+				throw new IllegalStateException(msg+", or set failOnWildcards=\"false\"");
+			}
+			log.warn(msg);
 			return true;
 		} 
 		throw new IllegalStateException("getBestMatchingElementPath unknown Term type ["+term.getClass().getName()+"]");
@@ -562,7 +588,15 @@ public abstract class ToXml<C,N> extends XmlAligner {
 			throw new SAXException(msg);
 		}		
 	}
-	
+
+	public void handleError(SAXException e) throws SAXException {
+		ErrorHandler errorHandler=validatorHandler.getErrorHandler();
+		if (errorHandler!=null) {
+			errorHandler.error(new SAXParseException(e.getMessage(),null));
+		} else {
+			throw e;
+		}		
+	}
 	
 	public String getQName(String namespace, String name) throws SAXException {
 		if (StringUtils.isNotEmpty(namespace)) {
@@ -662,6 +696,14 @@ public abstract class ToXml<C,N> extends XmlAligner {
 
 	public void setDeepSearch(boolean deepSearch) {
 		this.deepSearch = deepSearch;
+	}
+
+	public boolean isFailOnWildcards() {
+		return failOnWildcards;
+	}
+
+	public void setFailOnWildcards(boolean failOnWildcards) {
+		this.failOnWildcards = failOnWildcards;
 	}
 
 }